lambda-efs-cdk/requirements.txt
-e .
aws-cdk.core==1.117.0
aws-cdk.aws-apigatewayv2==1.117.0
aws-cdk.aws-apigatewayv2-integrations==1.117.0
aws-cdk.aws-ec2==1.117.0
aws-cdk.aws-lambda==1.117.0
aws-cdk.aws-efs==1.117.0